+++ Sources/Demo/DiscoverViewController.swift
@@ -1,17 +1,32 @@
+import MapboxMaps
 import MapboxSearch
+import MapboxSearchUI
 import MapKit
 import UIKit
 
 final class DiscoverViewController: UIViewController {
-    @IBOutlet private var mapView: MKMapView!
+    private var mapView = MapView(frame: .zero, mapInitOptions: defaultCameraOptions)
     @IBOutlet private var segmentedControl: UISegmentedControl!
 
     private let category = Discover()
+    lazy var annotationsManager = mapView.annotations.makePointAnnotationManager()
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        configureDefaultMapRegion()
+        mapView.translatesAutoresizingMaskIntoConstraints = false
+        view.addSubview(mapView)
+        view.sendSubviewToBack(mapView)
+
+        NSLayoutConstraint.activate([
+            mapView.topAnchor.constraint(equalTo: view.topAnchor),
+            mapView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            mapView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            mapView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+        ])
+
+        // Show user location
+        mapView.location.options.puckType = .puck2D()
     }
 }
 
@@ -20,12 +35,11 @@ final class DiscoverViewController: UIViewController {
 extension DiscoverViewController {
     @IBAction
     private func handleSearchInRegionAction() {
-        let regionResultsLimit: Int
-        switch category.apiType {
+        let regionResultsLimit = switch category.apiType {
         case .geocoding:
-            regionResultsLimit = 10
+            10
         default:
-            regionResultsLimit = 100
+            100
         }
 
         category.search(
@@ -47,15 +61,9 @@ extension DiscoverViewController {
 // MARK: - Private
 
 extension DiscoverViewController {
-    private var currentBoundingBox: BoundingBox {
-        let rect = mapView.visibleMapRect
-        let neMapPoint = MKMapPoint(x: rect.maxX, y: rect.origin.y)
-        let swMapPoint = MKMapPoint(x: rect.origin.x, y: rect.maxY)
-
-        let neCoordinate = neMapPoint.coordinate
-        let swCoordinate = swMapPoint.coordinate
-
-        return .init(swCoordinate, neCoordinate)
+    private var currentBoundingBox: MapboxSearch.BoundingBox {
+        let bounds = mapView.mapboxMap.coordinateBounds(for: mapView.bounds)
+        return MapboxSearch.BoundingBox(bounds.southwest, bounds.northeast)
     }
 
     private var currentSelectedCategory: Discover.Query {
@@ -68,28 +76,53 @@ extension DiscoverViewController {
         return allDemoCategories[segmentedControl.selectedSegmentIndex]
     }
 
-    private func configureDefaultMapRegion() {
-        let nyLocation = CLLocationCoordinate2D(latitude: 40.730610, longitude: -73.935242)
-        let span = MKCoordinateSpan(latitudeDelta: 0.15, longitudeDelta: 0.15)
-
-        let region = MKCoordinateRegion(
-            center: nyLocation,
-            span: span
+    private static var defaultCameraOptions: MapInitOptions {
+        let cameraOptions = CameraOptions(
+            center: CLLocationCoordinate2D(latitude: 40.730610, longitude: -73.935242),
+            zoom: 10.5
         )
-        mapView.setRegion(region, animated: false)
+
+        return MapInitOptions(cameraOptions: cameraOptions)
     }
 
-    private func showCategoryResults(_ results: [Discover.Result]) {
-        mapView.removeAnnotations(mapView.annotations)
+    private func cameraToAnnotations(_ annotations: [PointAnnotation]) {
+        if annotations.count == 1, let annotation = annotations.first {
+            mapView.camera.fly(
+                to: .init(center: annotation.point.coordinates, zoom: 15),
+                duration: 0.25,
+                completion: nil
+            )
+        } else {
+            let coordinatesCamera = mapView.mapboxMap.camera(
+                for: annotations.map(\.point.coordinates),
+                padding: UIEdgeInsets(
+                    top: 24,
+                    left: 24,
+                    bottom: 24,
+                    right: 24
+                ),
+                bearing: nil,
+                pitch: nil
+            )
+            mapView.camera.fly(to: coordinatesCamera, duration: 0.25, completion: nil)
+        }
+    }
 
-        let annotations: [MKPointAnnotation] = results.map {
-            let annotation = MKPointAnnotation()
-            annotation.coordinate = $0.coordinate
-            annotation.title = $0.name
+    private func showCategoryResults(_ results: [Discover.Result], cameraShouldFollow: Bool = true) {
+        annotationsManager.annotations = results.map {
+            var point = PointAnnotation(coordinate: $0.coordinate)
+            point.textField = $0.name
 
-            return annotation
+            /// Display a corresponding Maki icon for this Result when available
+            if let name = $0.makiIcon, let maki = Maki(rawValue: name) {
+                point.image = .init(image: maki.icon, name: maki.name)
+            }
+
+            return point
         }
 
-        mapView.showAnnotations(annotations, animated: true)
+        if cameraShouldFollow {
+            cameraToAnnotations(annotationsManager.annotations)
+        }
     }
 }